requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
datetime
plotly
pillow
altair<5